requirements.txt
Pillow
xmltodict
